+++ src/Node.h
@@ -33,12 +33,12 @@ public:
 protected:
     uint8_t investigateSlot(uint8_t x, uint8_t y, uint8_t dX, uint8_t dY, uint8_t piece);
 
-    static long int xCount;
-    static long int oCount;
-    static long int tieCount;
+    static long long unsigned int xCount;
+    static long long unsigned int oCount;
+    static long long unsigned int tieCount;
     uint8_t turn; //Whose turn it is RIGHT NOW. As in, ON THIS NODE!
 
-    static int count;
+    static long long unsigned int count;
 
     uint8_t board[GRID_X][GRID_Y];
 };
